requirements.txt
matplotlib==3.8.3
numpy==1.26.4
Pillow==9.5.0
Pillow==10.3.0
scikit_learn==1.5.0
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
